feat: add auto-mode quiz flow, fix WebSocket token refresh and scoring bug
- Implement auto-mode quiz flow (timer expiry → auto-advance → next/end)
- Add AutoModeControls, AutoAdvanceCountdown components
- Add useAutoAdvance hook with countdown logic
- Extend session store with auto-mode state and localStorage persistence
- Enhance TimerDisplay with questionId/sessionState guards
- Fix WebSocket reconnection: refresh JWT before reconnecting
- Fix scoring bug: multiply score by SCORE_MULTIPLIER for composite encoding
- Add property-based tests (6 properties, 100 runs each)
- Add unit and integration tests (545 tests passing)

+++ backend/session-service/src/main/java/com/quizplatform/session/service/RedisSessionService.java
@@ -295,7 +295,7 @@ public class RedisSessionService {
     /**
      * Update the leaderboard score for a participant using ZINCRBY.
      */
-    public void incrementLeaderboardScore(String pin, String participantId, int scoreToAdd) {
+    public void incrementLeaderboardScore(String pin, String participantId, long scoreToAdd) {
         String key = leaderboardKey(pin);
         redisTemplate.opsForZSet().incrementScore(key, participantId, scoreToAdd);
         redisTemplate.expire(key, Duration.ofSeconds(SESSION_TTL_SECONDS));
